Module google_bigtableadmin2::api

Structs

AppProfile
A configuration object describing how Cloud Bigtable should treat traffic from a particular end user application.

AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both allServices and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { "audit_configs": [ { "service": "allServices", "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" }, { "log_type": "ADMIN_READ" } ] }, { "service": "sampleservice.googleapis.com", "audit_log_configs": [ { "log_type": "DATA_READ" }, { "log_type": "DATA_WRITE", "exempted_members": [ "user:aliya@example.com" ] } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts jose@example.com from DATA_READ logging, and aliya@example.com from DATA_WRITE logging.

AuditLogConfig
Provides the configuration for logging a type of permissions. Example: { "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" } ] } This enables 'DATA_READ' and 'DATA_WRITE' logging, while exempting jose@example.com from DATA_READ logging.

Backup
A backup of a Cloud Bigtable table.

BackupInfo
Information about a backup.

BigtableAdmin
Central instance to access all BigtableAdmin related resource activities

Binding
Associates members with a role.

CheckConsistencyRequest
Request message for google.bigtable.admin.v2.BigtableTableAdmin.CheckConsistency

CheckConsistencyResponse
Response message for google.bigtable.admin.v2.BigtableTableAdmin.CheckConsistency

Cluster
A resizable group of nodes in a particular cloud location, capable of serving all Tables in the parent Instance.

ClusterState
The state of a table's data in a particular cluster.

ColumnFamily
A set of columns within a table which share a common configuration.

CreateInstanceRequest
Request message for BigtableInstanceAdmin.CreateInstance.

CreateTableRequest
Request message for google.bigtable.admin.v2.BigtableTableAdmin.CreateTable

DropRowRangeRequest
Request message for google.bigtable.admin.v2.BigtableTableAdmin.DropRowRange

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EncryptionConfig
Cloud Key Management Service (Cloud KMS) settings for a CMEK-protected cluster.

EncryptionInfo
Encryption information for a given resource. If this resource is protected with customer managed encryption, the in-use Cloud Key Management Service (Cloud KMS) key version is specified along with its status.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

GcRule
Rule for determining which cells to delete during garbage collection.

GenerateConsistencyTokenRequest
Request message for google.bigtable.admin.v2.BigtableTableAdmin.GenerateConsistencyToken

GenerateConsistencyTokenResponse
Response message for google.bigtable.admin.v2.BigtableTableAdmin.GenerateConsistencyToken

GetIamPolicyRequest
Request message for GetIamPolicy method.

GetPolicyOptions
Encapsulates settings provided to GetIamPolicy.

Instance
A collection of Bigtable Tables and the resources that serve them. All tables in an instance are served from all Clusters in the instance.

Intersection
A GcRule which deletes cells matching all of the given rules.

ListAppProfilesResponse
Response message for BigtableInstanceAdmin.ListAppProfiles.

ListBackupsResponse
The response for ListBackups.

ListClustersResponse
Response message for BigtableInstanceAdmin.ListClusters.

ListInstancesResponse
Response message for BigtableInstanceAdmin.ListInstances.

ListLocationsResponse
The response message for Locations.ListLocations.

ListOperationsResponse
The response message for Operations.ListOperations.

ListTablesResponse
Response message for google.bigtable.admin.v2.BigtableTableAdmin.ListTables

Location
A resource that represents Google Cloud Platform location.

Modification
A create, update, or delete of a particular column family.

ModifyColumnFamiliesRequest
Request message for google.bigtable.admin.v2.BigtableTableAdmin.ModifyColumnFamilies

MultiClusterRoutingUseAny
Read/write requests are routed to the nearest cluster in the instance, and will fail over to the nearest cluster that is available in the event of transient errors or delays. Clusters in a region are considered equidistant. Choosing this option sacrifices read-your-writes consistency to improve availability.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

OperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the BigtableAdmin hub.

OperationProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

ProjectInstanceAppProfileCreateCall
Creates an app profile within an instance.

ProjectInstanceAppProfileDeleteCall
Deletes an app profile from an instance.

ProjectInstanceAppProfileGetCall
Gets information about an app profile.

ProjectInstanceAppProfileListCall
Lists information about app profiles in an instance.

ProjectInstanceAppProfilePatchCall
Updates an app profile within an instance.

ProjectInstanceClusterBackupCreateCall
Starts creating a new Cloud Bigtable Backup. The returned backup long-running operation can be used to track creation of the backup. The metadata field type is CreateBackupMetadata. The response field type is Backup, if successful. Cancelling the returned operation will stop the creation and delete the backup.

ProjectInstanceClusterBackupDeleteCall
Deletes a pending or completed Cloud Bigtable backup.

ProjectInstanceClusterBackupGetCall
Gets metadata on a pending or completed Cloud Bigtable Backup.

ProjectInstanceClusterBackupGetIamPolicyCall
Gets the access control policy for a Table resource. Returns an empty policy if the resource exists but does not have a policy set.

ProjectInstanceClusterBackupListCall
Lists Cloud Bigtable backups. Returns both completed and pending backups.

ProjectInstanceClusterBackupPatchCall
Updates a pending or completed Cloud Bigtable Backup.

ProjectInstanceClusterBackupSetIamPolicyCall
Sets the access control policy on a Table resource. Replaces any existing policy.

ProjectInstanceClusterBackupTestIamPermissionCall
Returns permissions that the caller has on the specified table resource.

ProjectInstanceClusterCreateCall
Creates a cluster within an instance.

ProjectInstanceClusterDeleteCall
Deletes a cluster from an instance.

ProjectInstanceClusterGetCall
Gets information about a cluster.

ProjectInstanceClusterListCall
Lists information about clusters in an instance.

ProjectInstanceClusterUpdateCall
Updates a cluster within an instance.

ProjectInstanceCreateCall
Create an instance within a project.

ProjectInstanceDeleteCall
Delete an instance from a project.

ProjectInstanceGetCall
Gets information about an instance.

ProjectInstanceGetIamPolicyCall
Gets the access control policy for an instance resource. Returns an empty policy if an instance exists but does not have a policy set.

ProjectInstanceListCall
Lists information about instances in a project.

ProjectInstancePartialUpdateInstanceCall
Partially updates an instance within a project. This method can modify all fields of an Instance and is the preferred way to update an Instance.

ProjectInstanceSetIamPolicyCall
Sets the access control policy on an instance resource. Replaces any existing policy.

ProjectInstanceTableCheckConsistencyCall
Checks replication consistency based on a consistency token, that is, if replication has caught up based on the conditions specified in the token and the check request.

ProjectInstanceTableCreateCall
Creates a new table in the specified instance. The table can be created with a full set of initial column families, specified in the request.

ProjectInstanceTableDeleteCall
Permanently deletes a specified table and all of its data.

ProjectInstanceTableDropRowRangeCall
Permanently drop/delete a row range from a specified table. The request can specify whether to delete all rows in a table, or only those that match a particular prefix.

ProjectInstanceTableGenerateConsistencyTokenCall
Generates a consistency token for a Table, which can be used in CheckConsistency to check whether mutations to the table that finished before this call started have been replicated. The tokens will be available for 90 days.

ProjectInstanceTableGetCall
Gets metadata information about the specified table.

ProjectInstanceTableGetIamPolicyCall
Gets the access control policy for a Table resource. Returns an empty policy if the resource exists but does not have a policy set.

ProjectInstanceTableListCall
Lists all tables served from a specified instance.

ProjectInstanceTableModifyColumnFamilyCall
Performs a series of column family modifications on the specified table. Either all or none of the modifications will occur before this method returns, but data requests received prior to that point may see a table where only some modifications have taken effect.

ProjectInstanceTableRestoreCall
Create a new table by restoring from a completed backup. The new table must be in the same instance as the instance containing the backup. The returned table long-running operation can be used to track the progress of the operation, and to cancel it. The metadata field type is RestoreTableMetadata. The response type is Table, if successful.

ProjectInstanceTableSetIamPolicyCall
Sets the access control policy on a Table resource. Replaces any existing policy.

ProjectInstanceTableTestIamPermissionCall
Returns permissions that the caller has on the specified table resource.

ProjectInstanceTestIamPermissionCall
Returns permissions that the caller has on the specified instance resource.

ProjectInstanceUpdateCall
Updates an instance within a project. This method updates only the display name and type for an Instance. To update other Instance properties, such as labels, use PartialUpdateInstance.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the BigtableAdmin hub.

RestoreInfo
Information about a table restore.

RestoreTableRequest
The request for RestoreTable.

SetIamPolicyRequest
Request message for SetIamPolicy method.

SingleClusterRouting
Unconditionally routes all read/write requests to a specific cluster. This option preserves read-your-writes consistency but does not improve availability.

Split
An initial split point for a newly created table.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

Table
A collection of user data indexed by row, column, and timestamp. Each table is served using the resources of its parent cluster.

TestIamPermissionsRequest
Request message for TestIamPermissions method.

TestIamPermissionsResponse
Response message for TestIamPermissions method.

Union
A GcRule which deletes cells matching any of the given rules.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.