Struct google_bigtableadmin2::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the BigtableAdmin hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_bigtableadmin2 as bigtableadmin2;
 
use std::default::Default;
use oauth2;
use bigtableadmin2::BigtableAdmin;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = BigtableAdmin::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `instances_app_profiles_create(...)`, `instances_app_profiles_delete(...)`, `instances_app_profiles_get(...)`, `instances_app_profiles_list(...)`, `instances_app_profiles_patch(...)`, `instances_clusters_backups_create(...)`, `instances_clusters_backups_delete(...)`, `instances_clusters_backups_get(...)`, `instances_clusters_backups_get_iam_policy(...)`, `instances_clusters_backups_list(...)`, `instances_clusters_backups_patch(...)`, `instances_clusters_backups_set_iam_policy(...)`, `instances_clusters_backups_test_iam_permissions(...)`, `instances_clusters_create(...)`, `instances_clusters_delete(...)`, `instances_clusters_get(...)`, `instances_clusters_list(...)`, `instances_clusters_update(...)`, `instances_create(...)`, `instances_delete(...)`, `instances_get(...)`, `instances_get_iam_policy(...)`, `instances_list(...)`, `instances_partial_update_instance(...)`, `instances_set_iam_policy(...)`, `instances_tables_check_consistency(...)`, `instances_tables_create(...)`, `instances_tables_delete(...)`, `instances_tables_drop_row_range(...)`, `instances_tables_generate_consistency_token(...)`, `instances_tables_get(...)`, `instances_tables_get_iam_policy(...)`, `instances_tables_list(...)`, `instances_tables_modify_column_families(...)`, `instances_tables_restore(...)`, `instances_tables_set_iam_policy(...)`, `instances_tables_test_iam_permissions(...)`, `instances_test_iam_permissions(...)`, `instances_update(...)`, `locations_get(...)` and `locations_list(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn instances_app_profiles_create(
&self,
request: AppProfile,
parent: &str
) -> ProjectInstanceAppProfileCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an app profile within an instance.

Arguments

request - No description provided.

parent - Required. The unique name of the instance in which to create the new app profile. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_app_profiles_delete(
&self,
name: &str
) -> ProjectInstanceAppProfileDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an app profile from an instance.

Arguments

name - Required. The unique name of the app profile to be deleted. Values are of the form projects/{project}/instances/{instance}/appProfiles/{app_profile}.

pub fn instances_app_profiles_get(
&self,
name: &str
) -> ProjectInstanceAppProfileGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about an app profile.

Arguments

name - Required. The unique name of the requested app profile. Values are of the form projects/{project}/instances/{instance}/appProfiles/{app_profile}.

pub fn instances_app_profiles_list(
&self,
parent: &str
) -> ProjectInstanceAppProfileListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about app profiles in an instance.

Arguments

parent - Required. The unique name of the instance for which a list of app profiles is requested. Values are of the form projects/{project}/instances/{instance}. Use {instance} = '-' to list AppProfiles for all Instances in a project, e.g., projects/myproject/instances/-.

pub fn instances_app_profiles_patch(
&self,
request: AppProfile,
name: &str
) -> ProjectInstanceAppProfilePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an app profile within an instance.

Arguments

request - No description provided.

name - The unique name of the app profile. Values are of the form projects/{project}/instances/{instance}/appProfiles/_a-zA-Z0-9*.

pub fn instances_clusters_backups_create(
&self,
request: Backup,
parent: &str
) -> ProjectInstanceClusterBackupCreateCall<'a, C>

Create a builder to help you perform the following task:

Starts creating a new Cloud Bigtable Backup. The returned backup long-running operation can be used to track creation of the backup. The metadata field type is CreateBackupMetadata. The response field type is Backup, if successful. Cancelling the returned operation will stop the creation and delete the backup.

Arguments

request - No description provided.

parent - Required. This must be one of the clusters in the instance in which this table is located. The backup will be stored in this cluster. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}.

pub fn instances_clusters_backups_delete(
&self,
name: &str
) -> ProjectInstanceClusterBackupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a pending or completed Cloud Bigtable backup.

Arguments

name - Required. Name of the backup to delete. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}/backups/{backup}.

pub fn instances_clusters_backups_get(
&self,
name: &str
) -> ProjectInstanceClusterBackupGetCall<'a, C>

Create a builder to help you perform the following task:

Gets metadata on a pending or completed Cloud Bigtable Backup.

Arguments

name - Required. Name of the backup. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}/backups/{backup}.

pub fn instances_clusters_backups_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectInstanceClusterBackupGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a Table resource. Returns an empty policy if the resource exists but does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_clusters_backups_list(
&self,
parent: &str
) -> ProjectInstanceClusterBackupListCall<'a, C>

Create a builder to help you perform the following task:

Lists Cloud Bigtable backups. Returns both completed and pending backups.

Arguments

parent - Required. The cluster to list backups from. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}. Use {cluster} = '-' to list backups for all clusters in an instance, e.g., projects/{project}/instances/{instance}/clusters/-.

pub fn instances_clusters_backups_patch(
&self,
request: Backup,
name: &str
) -> ProjectInstanceClusterBackupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a pending or completed Cloud Bigtable Backup.

Arguments

request - No description provided.

name - A globally unique identifier for the backup which cannot be changed. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}/ backups/_a-zA-Z0-9* The final segment of the name must be between 1 and 50 characters in length. The backup is stored in the cluster identified by the prefix of the backup name of the form projects/{project}/instances/{instance}/clusters/{cluster}.

pub fn instances_clusters_backups_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectInstanceClusterBackupSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on a Table resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn instances_clusters_backups_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectInstanceClusterBackupTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that the caller has on the specified table resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_clusters_create(
&self,
request: Cluster,
parent: &str
) -> ProjectInstanceClusterCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a cluster within an instance.

Arguments

request - No description provided.

parent - Required. The unique name of the instance in which to create the new cluster. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_clusters_delete(
&self,
name: &str
) -> ProjectInstanceClusterDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a cluster from an instance.

Arguments

name - Required. The unique name of the cluster to be deleted. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}.

pub fn instances_clusters_get(
&self,
name: &str
) -> ProjectInstanceClusterGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a cluster.

Arguments

name - Required. The unique name of the requested cluster. Values are of the form projects/{project}/instances/{instance}/clusters/{cluster}.

pub fn instances_clusters_list(
&self,
parent: &str
) -> ProjectInstanceClusterListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about clusters in an instance.

Arguments

parent - Required. The unique name of the instance for which a list of clusters is requested. Values are of the form projects/{project}/instances/{instance}. Use {instance} = '-' to list Clusters for all Instances in a project, e.g., projects/myproject/instances/-.

pub fn instances_clusters_update(
&self,
request: Cluster,
name: &str
) -> ProjectInstanceClusterUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates a cluster within an instance.

Arguments

request - No description provided.

name - The unique name of the cluster. Values are of the form projects/{project}/instances/{instance}/clusters/a-z*.

pub fn instances_tables_check_consistency(
&self,
request: CheckConsistencyRequest,
name: &str
) -> ProjectInstanceTableCheckConsistencyCall<'a, C>

Create a builder to help you perform the following task:

Checks replication consistency based on a consistency token, that is, if replication has caught up based on the conditions specified in the token and the check request.

Arguments

request - No description provided.

name - Required. The unique name of the Table for which to check replication consistency. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_create(
&self,
request: CreateTableRequest,
parent: &str
) -> ProjectInstanceTableCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new table in the specified instance. The table can be created with a full set of initial column families, specified in the request.

Arguments

request - No description provided.

parent - Required. The unique name of the instance in which to create the table. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_tables_delete(
&self,
name: &str
) -> ProjectInstanceTableDeleteCall<'a, C>

Create a builder to help you perform the following task:

Permanently deletes a specified table and all of its data.

Arguments

name - Required. The unique name of the table to be deleted. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_drop_row_range(
&self,
request: DropRowRangeRequest,
name: &str
) -> ProjectInstanceTableDropRowRangeCall<'a, C>

Create a builder to help you perform the following task:

Permanently drop/delete a row range from a specified table. The request can specify whether to delete all rows in a table, or only those that match a particular prefix.

Arguments

request - No description provided.

name - Required. The unique name of the table on which to drop a range of rows. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_generate_consistency_token(
&self,
request: GenerateConsistencyTokenRequest,
name: &str
) -> ProjectInstanceTableGenerateConsistencyTokenCall<'a, C>

Create a builder to help you perform the following task:

Generates a consistency token for a Table, which can be used in CheckConsistency to check whether mutations to the table that finished before this call started have been replicated. The tokens will be available for 90 days.

Arguments

request - No description provided.

name - Required. The unique name of the Table for which to create a consistency token. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_get(
&self,
name: &str
) -> ProjectInstanceTableGetCall<'a, C>

Create a builder to help you perform the following task:

Gets metadata information about the specified table.

Arguments

name - Required. The unique name of the requested table. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectInstanceTableGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a Table resource. Returns an empty policy if the resource exists but does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_tables_list(
&self,
parent: &str
) -> ProjectInstanceTableListCall<'a, C>

Create a builder to help you perform the following task:

Lists all tables served from a specified instance.

Arguments

parent - Required. The unique name of the instance for which tables should be listed. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_tables_modify_column_families(
&self,
request: ModifyColumnFamiliesRequest,
name: &str
) -> ProjectInstanceTableModifyColumnFamilyCall<'a, C>

Create a builder to help you perform the following task:

Performs a series of column family modifications on the specified table. Either all or none of the modifications will occur before this method returns, but data requests received prior to that point may see a table where only some modifications have taken effect.

Arguments

request - No description provided.

name - Required. The unique name of the table whose families should be modified. Values are of the form projects/{project}/instances/{instance}/tables/{table}.

pub fn instances_tables_restore(
&self,
request: RestoreTableRequest,
parent: &str
) -> ProjectInstanceTableRestoreCall<'a, C>

Create a builder to help you perform the following task:

Create a new table by restoring from a completed backup. The new table must be in the same instance as the instance containing the backup. The returned table long-running operation can be used to track the progress of the operation, and to cancel it. The metadata field type is RestoreTableMetadata. The response type is Table, if successful.

Arguments

request - No description provided.

parent - Required. The name of the instance in which to create the restored table. This instance must be the parent of the source backup. Values are of the form projects//instances/.

pub fn instances_tables_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectInstanceTableSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on a Table resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn instances_tables_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectInstanceTableTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that the caller has on the specified table resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_create(
&self,
request: CreateInstanceRequest,
parent: &str
) -> ProjectInstanceCreateCall<'a, C>

Create a builder to help you perform the following task:

Create an instance within a project.

Arguments

request - No description provided.

parent - Required. The unique name of the project in which to create the new instance. Values are of the form projects/{project}.

pub fn instances_delete(&self, name: &str) -> ProjectInstanceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Delete an instance from a project.

Arguments

name - Required. The unique name of the instance to be deleted. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_get(&self, name: &str) -> ProjectInstanceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about an instance.

Arguments

name - Required. The unique name of the requested instance. Values are of the form projects/{project}/instances/{instance}.

pub fn instances_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectInstanceGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for an instance resource. Returns an empty policy if an instance exists but does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_list(&self, parent: &str) -> ProjectInstanceListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about instances in a project.

Arguments

parent - Required. The unique name of the project for which a list of instances is requested. Values are of the form projects/{project}.

pub fn instances_partial_update_instance(
&self,
request: Instance,
name: &str
) -> ProjectInstancePartialUpdateInstanceCall<'a, C>

Create a builder to help you perform the following task:

Partially updates an instance within a project. This method can modify all fields of an Instance and is the preferred way to update an Instance.

Arguments

request - No description provided.

name - The unique name of the instance. Values are of the form projects/{project}/instances/a-z+[a-z0-9].

pub fn instances_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectInstanceSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on an instance resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn instances_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectInstanceTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that the caller has on the specified instance resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn instances_update(
&self,
request: Instance,
name: &str
) -> ProjectInstanceUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates an instance within a project. This method updates only the display name and type for an Instance. To update other Instance properties, such as labels, use PartialUpdateInstance.

Arguments

request - No description provided.

name - The unique name of the instance. Values are of the form projects/{project}/instances/a-z+[a-z0-9].

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.